+++ lib/rb-youtube-dl/options.rb
