+++ lib/rb-youtube-dl/support.rb
